requirements.txt
Scrapy==2.11.2
psycopg2-binary==2.8.6
pymongo
redis
